+++ CHANGELOG.md
@@ -7,6 +7,12 @@
   - Nuvlabox status - Update heartbeat even if NuvlaBox encounter a spec
     validation issue
 
+## [5.19.1] - 2021-11-10
+
+### Changed
+
+  - Pricing regression fix
+
 ## [5.19.0] - 2021-11-08
 
 ### Added